Started to create the GUI
I created an sketch of what the final product will most likely be able to do.
The GUI here is most probably not the final one.
I have the city map and the edit buttons; each button goes to a edit window. In the new edit windows I can Create Delete and Update; The Read I want it from the main GUI